refactor: move TudeeAppScaffold to another package
