+++ app/src/main/java/com/example/tudeeapp/presentation/screen/taskManagement/TaskManagementBottomSheet.kt
@@ -12,7 +12,7 @@ import com.example.tudeeapp.R
 import com.example.tudeeapp.presentation.common.components.TudeeBottomSheet
 import com.example.tudeeapp.presentation.common.components.TudeeDatePickerDialog
 import com.example.tudeeapp.presentation.navigation.LocalNavController
-import com.example.tudeeapp.presentation.navigation.LocalSnackBarState
+import com.example.tudeeapp.presentation.LocalSnackBarState
 import com.example.tudeeapp.presentation.screen.taskManagement.components.CategoryGrid
 import com.example.tudeeapp.presentation.screen.taskManagement.components.PriorityRow
 import com.example.tudeeapp.presentation.screen.taskManagement.components.TaskManagementButtons
